requirements.txt
streamlit==0.82.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0